ss10 module2

+++ modele_2/src/ss1_introduction_to_java/bai_tap/ChuyenDoiTienTe.java
@@ -4,11 +4,11 @@ import java.util.Scanner;
 
 public class ChuyenDoiTienTe {
     public static void main(String[] args) {
-        int rate = 23000;
+        double rate = 23000;
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter currency dola");
-        int dola = scanner.nextInt();
-        int vnd = dola*rate;
+        double dola = scanner.nextDouble();
+        double vnd = dola*rate;
         System.out.println("Convert to VND: "+ vnd+" VND");
 
     }
